Widget anchored at top right corner

+++ Sources/Core/RMath.cpp
@@ -155,24 +155,24 @@ namespace At0::Ray
 	}
 	RAY_EXPORT Float2 ScreenSpaceToNDCSpace(Float2 coords)
 	{
-		return { coords.x / (0.5f * Window::Get().GetFramebufferSize().x) - 1,
-			coords.y / (0.5f * Window::Get().GetFramebufferSize().y) - 1 };
+		return { coords.x / (0.5f * (float)Window::Get().GetFramebufferSize().x) - 1,
+			coords.y / (0.5f * (float)Window::Get().GetFramebufferSize().y) - 1 };
 	}
 	RAY_EXPORT float NDCSpaceToScreenSpaceX(float x)
 	{
-		return (x + 1) * 0.5f * Window::Get().GetFramebufferSize().x;
+		return (x + 1) * 0.5f * (float)Window::Get().GetFramebufferSize().x;
 	}
 	RAY_EXPORT float ScreenSpaceToNDCSpaceX(float x)
 	{
-		return x / (0.5f * Window::Get().GetFramebufferSize().x) - 1;
+		return x / (0.5f * (float)Window::Get().GetFramebufferSize().x) - 1;
 	}
 	RAY_EXPORT float NDCSpaceToScreenSpaceY(float y)
 	{
-		return (y + 1) * 0.5f * Window::Get().GetFramebufferSize().y;
+		return (y + 1) * 0.5f * (float)Window::Get().GetFramebufferSize().y;
 	}
 	RAY_EXPORT float ScreenSpaceToNDCSpaceY(float y)
 	{
-		return y / (0.5f * Window::Get().GetFramebufferSize().y) - 1;
+		return y / (0.5f * (float)Window::Get().GetFramebufferSize().y) - 1;
 	}
 
 	inline std::ostream& operator<<(std::ostream& os, const Int2& data)
